fix: Update Dockerfile to run as root
The Dockerfile has been updated to change the user running the container from "nextjs" to "root". This change is necessary to allow the container to bind to port 80, which requires root privileges. By running the container as root, the application will be able to listen on port 80 and serve incoming requests.