refactor: Update ViewPage component to use parsedTags in getGallery function
The ViewPage component has been refactored to use the parsedTags variable in the getGallery function. This change ensures that the tags are correctly passed to the saveImage function when saving an image. It improves the functionality of the ViewPage component by ensuring that the correct tags are associated with each image.