refactor: Update ViewPage component to handle selected tags and show all images
The ViewPage component in the viewpage.tsx file has been refactored to handle selected tags and show all images. This change includes updating the handleSelectAllTags function to set the selected tags and show all images, improving the functionality of the component. It ensures that when all tags are selected, all images are displayed. This enhancement enhances the user experience of the ViewPage component.