refactor: Add selectedImageId state and update tag update API endpoint
The EditPage component in edit.tsx has been updated to include a new state variable called selectedImageId. This state variable is used to keep track of the currently selected image ID. Additionally, the API endpoint for updating tags has been modified in the route.ts file to include the image ID in the request body. These changes improve the functionality of the EditPage component by allowing the user to select an image and update its tags.