refactor: Add GET route for retrieving public galleries
A new GET route has been added to the route.ts file. This route retrieves public galleries and returns them as a JSON response. It utilizes the getPublicGalleries function from the galleries module to fetch the galleries data. This addition enhances the functionality of the API by allowing clients to access public galleries easily.